Extract shared navigation logic into AbstractSyntaxTreeNodesMotionHandler and update handlers to extend it
