+++ src/main/java/com/zyurkalov/ideavim/syntaxtreejumper/motions/ExpressionMotionHandler.java
@@ -10,7 +10,7 @@ import com.zyurkalov.ideavim.syntaxtreejumper.adapters.SyntaxTreeAdapter;
  * in accordance to the given Direction from the caret, then places the caret
  * at the found element.
  */
-public class ExpressionMotionHandler extends SyntaxTreeNodesMotionHandler {
+public class ExpressionMotionHandler extends AbstractSyntaxTreeNodesMotionHandler {
 
     public ExpressionMotionHandler(SyntaxTreeAdapter syntaxTree, MotionDirection direction) {
         super(syntaxTree, direction);
