+++ src/test/java/com/zyurkalov/ideavim/syntaxtreejumper/motions/ShrinkExpandMotionHandlerTest.java
@@ -202,7 +202,7 @@ class ShrinkExpandMotionHandlerTest {
         }
 
         PsiFile javaFile = myFixture.configureByText("TestClass.java", javaCode);
-        SyntaxTreeNodesMotionHandler handler = new SyntaxTreeNodesMotionHandler( new PsiSyntaxTreeAdapter(javaFile), testData.motionType);
+        AbstractSyntaxTreeNodesMotionHandler handler = new SyntaxTreeNodesMotionHandler( new PsiSyntaxTreeAdapter(javaFile), testData.motionType);
 
         ApplicationManager.getApplication().runReadAction(() -> {
             Optional<Offsets> result = handler.findNext(testData.initialOffsets);
